Decode in chunks for edge-case documents in 32 bit
Decoding big openjpeg documents, say in sizes 12608x16806 and above, hit
calculations that cannot be stored in the 32bit WASM build using
straight opj_decode. To fit the decoding into the 32bit WASM runtime
limits we split the document in height-strips of 256 and decode the
document chunks.

To decode the images, we always convert the image to 32bpp RGBA. This
may not be compatible with PDF.js itself.